Move ClusterTPAssociationProducer declaration to the source file
